+++ CHANGELOG.md
@@ -252,3 +252,4 @@
 * Define `Equivalence` datatype and define `Contravariant` instance for it
 * Define `swop` for `Wye` (useful if you want to invert binary tree with `hoist swap`)
 * Add experimental `Monotonic` typeclass
+* Implement delete method for Stack